+++ config/config.exs
@@ -2,33 +2,6 @@
 # and its dependencies with the aid of the Mix.Config module.
 use Mix.Config
 
-# This configuration is loaded before any dependency and is restricted
-# to this project. If another project depends on this project, this
-# file won't be loaded nor affect the parent project. For this reason,
-# if you want to provide default values for your application for
-# 3rd-party users, it should be done in your "mix.exs" file.
-
-# You can configure for your application as:
-#
-#     config :sirko, key: :value
-#
-# And access this configuration in your application as:
-#
-#     Application.get_env(:sirko, :key)
-#
-# Or configure a 3rd-party app:
-#
-#     config :logger, level: :info
-#
-
-# It is also possible to import configuration files, relative to this
-# directory. For example, you can emulate configuration per environment
-# by uncommenting the line below and defining dev.exs, test.exs and such.
-# Configuration from the imported file will override the ones defined
-# here (which is why it is important to import them last).
-#
-#     import_config "#{Mix.env}.exs"
-
 config :bolt_sips, Bolt,
   url: System.get_env("NEO4J_URL") || "bolt://localhost:7687",
   ssl: System.get_env("NEO4J_SSL") || false,
@@ -40,9 +13,16 @@ config :sirko, :web,
   port: System.get_env("SIRKO_HTTP_PORT") || 4000,
   client_url: System.get_env("SIRKO_CLIENT_URL") # the address of a site for which predictions should be made
 
+# Settings for sort of an internal CRON (https://en.wikipedia.org/wiki/Cron)
 config :sirko, :scheduler,
-  expire_sessions_in: 3600 * 1000, # how often the scheduler should be launched to expire inactive sessions
-  remove_stale_data_in: 3600 * 1000 * 24 # how often the scheduler should be launched to remove stale data
+  expire_sessions_every: 3600 * 1000, # how often the scheduler should be launched to expire inactive sessions
+  remove_stale_data_every: 3600 * 1000 * 24 # how often the scheduler should be launched to remove stale data
+
+# Settings which affect the behavior of the engine. These settings are used by different parts of the engine.
+# To get more details about them, read the config schema: config/sirko.schema.exs
+config :sirko, :engine,
+  inactive_session_in: 3600 * 1000,
+  stale_session_in: 3600 * 1000 * 24 * 7
 
 config :logger, :console,
   level: (System.get_env("SIRKO_DEBUG_LEVEL") || "info") |> String.to_atom,