+++ lib/sirko/plugs/cors.ex
@@ -11,10 +11,11 @@ defmodule Sirko.Plugs.Cors do
 
   def call(conn, _ \\ []) do
     # TODO: find a better way to pass parameters to the plug.
-    opts = Application.get_env(:sirko, :web)
+    client_url = Application.get_env(:sirko, :web)
+    |> Keyword.get(:client_url)
 
     conn
-    |> put_resp_header("access-control-allow-origin", Keyword.get(opts, :client_url))
+    |> put_resp_header("access-control-allow-origin", client_url)
     |> put_resp_header("access-control-allow-methods", @allowed_http_methods)
     |> put_resp_header("access-control-allow-credentials", "true")
   end