+++ lib/sirko/scheduler/supervisor.ex
@@ -18,42 +18,46 @@ defmodule Sirko.Scheduler.Supervisor do
   end
 
   def init(opts) do
+    engine_opts = Application.get_env(:sirko, :engine)
+
     children = [
-      inactive_sessions_worker(opts),
-      stale_data_worker(opts)
+      inactive_sessions_worker(opts, engine_opts),
+      stale_data_worker(opts, engine_opts)
     ]
 
     supervise(children, strategy: :one_for_one)
   end
 
-  defp inactive_sessions_worker(opts) do
-    expire_sessions_in = Keyword.get(opts, :expire_sessions_in)
+  defp inactive_sessions_worker(worker_opts, engine_opts) do
+    expire_sessions_every = Keyword.get(worker_opts, :expire_sessions_every)
+    inactive_session_in = Keyword.get(engine_opts, :inactive_session_in)
 
     callback = fn ->
-      Sirko.Session.expire_all_inactive
+      Sirko.Session.expire_all_inactive(inactive_session_in)
     end
 
     name = Sirko.Scheduler.InactiveSessions
 
     worker(
       Worker,
-      [[timeout: expire_sessions_in, callback: callback, name: name]],
+      [[timeout: expire_sessions_every, callback: callback, name: name]],
       id: name
     )
   end
 
-  defp stale_data_worker(opts) do
-    remove_stale_data_in = Keyword.get(opts, :remove_stale_data_in)
+  defp stale_data_worker(worker_opts, engine_opts) do
+    remove_stale_data_every = Keyword.get(worker_opts, :remove_stale_data_every)
+    stale_session_in = Keyword.get(engine_opts, :stale_session_in)
 
     callback = fn ->
-      Sirko.Cleaner.clean_up
+      Sirko.Cleaner.clean_up(stale_session_in)
     end
 
     name = Sirko.Scheduler.Cleaner
 
     worker(
       Worker,
-      [[timeout: remove_stale_data_in, callback: callback, name: name]],
+      [[timeout: remove_stale_data_every, callback: callback, name: name]],
       id: name
     )
   end